Make CouponService internal to be only used as a helper within Yosemite
